refactor: make join, words and truncate tree shakable
