+++ src/_math.ts
