+++ src/primitives.ts
@@ -1,5 +1,6 @@
-import type { Math } from './math'
+import type { Math } from './_math'
 import type { TupleOf } from './internals'
+import { type Join, join } from './join'
 
 /**
  * Gets the character at the given index.
@@ -69,39 +70,6 @@ function endsWith<
   P extends number = Length<T>,
 >(text: T, search: S, position = text.length as P) {
   return text.endsWith(search, position) as EndsWith<T, S, P>
-}
-
-/**
- * Joins a tuple of strings with the given delimiter.
- * T: The tuple of strings to join.
- * delimiter: The delimiter.
- */
-type Join<
-  T extends readonly string[],
-  delimiter extends string = '',
-> = string[] extends T
-  ? string // Avoid spending resources on a wide type
-  : T extends readonly [
-      infer first extends string,
-      ...infer rest extends string[],
-    ]
-  ? rest extends []
-    ? first
-    : `${first}${delimiter}${Join<rest, delimiter>}`
-  : ''
-
-/**
- * A strongly-typed version of `Array.prototype.join`.
- * @param tuple the tuple of strings to join.
- * @param delimiter the delimiter.
- * @returns the joined string in both type level and runtime.
- * @example join(['hello', 'world'], '-') // 'hello-world'
- */
-function join<const T extends readonly string[], D extends string = ''>(
-  tuple: T,
-  delimiter: D = '' as D,
-) {
-  return tuple.join(delimiter) as Join<T, D>
 }
 
 /**
@@ -393,7 +361,6 @@ export type {
   Concat,
   EndsWith,
   Includes,
-  Join,
   Length,
   PadEnd,
   PadStart,
@@ -410,7 +377,6 @@ export {
   concat,
   endsWith,
   includes,
-  join,
   length,
   padEnd,
   padStart,
