+++ src/words.test.ts
@@ -1,4 +1,4 @@
-import * as subject from './utils'
+import * as subject from './words'
 
 type Mutable<Type> = {
   -readonly [Key in keyof Type]: Type[Key]
@@ -60,42 +60,5 @@ describe('words', () => {
     const result = subject.words(' someWeird-cased$*String1986Foo Bar ')
     expect(result).toEqual(expected)
     type test = Expect<Equal<typeof result, Mutable<typeof expected>>>
-  })
-})
-
-describe('truncate', () => {
-  test('truncate small sentence does nothing', () => {
-    const expected = 'Hello' as const
-    const result = subject.truncate('Hello', 9)
-    expect(result).toEqual(expected)
-    type test = Expect<Equal<typeof result, typeof expected>>
-  })
-
-  test('truncate big sentence truncate', () => {
-    const expected = 'Hello ...' as const
-    const result = subject.truncate('Hello world', 9)
-    expect(result).toEqual(expected)
-    type test = Expect<Equal<typeof result, typeof expected>>
-  })
-
-  test('truncate with negative integer does truncate', () => {
-    const expected = '...' as const
-    const result = subject.truncate('Hello world', -1)
-    expect(result).toEqual(expected)
-    type test = Expect<Equal<typeof result, typeof expected>>
-  })
-
-  test('truncate big sentence with specified omission', () => {
-    const expected = 'Hello[...]' as const
-    const result = subject.truncate('Hello world', 10, '[...]')
-    expect(result).toEqual(expected)
-    type test = Expect<Equal<typeof result, typeof expected>>
-  })
-
-  test('truncate small sentence with specified omission', () => {
-    const expected = 'Hello' as const
-    const result = subject.truncate('Hello', 10, '[...]')
-    expect(result).toEqual(expected)
-    type test = Expect<Equal<typeof result, typeof expected>>
   })
 })
